app/requirements.txt
annotated-types        0.7.0
bcrypt                 3.2.0
certifi                2024.8.30
cffi                   1.17.1
charset-normalizer     3.4.0
click                  8.1.7
colorama               0.4.6
databases              0.9.0
dnspython              2.7.0
email_validator        2.2.0
fastapi                0.115.4
fastapi-cli            0.0.5
greenlet               3.1.1
h11                    0.14.0
httpcore               1.0.6
httptools              0.6.4
httpx                  0.27.2
idna                   3.10
itsdangerous           2.2.0
Jinja2                 3.1.4
Mako                   1.3.6
markdown-it-py         3.0.0
MarkupSafe             3.0.2
mdurl                  0.1.2
mysql-connector-python 9.1.0
mysqlclient            2.2.5
orjson                 3.10.11
passlib                1.7.4
pip                    24.3.1
pycparser              2.22
pycryptodome           3.21.0
pydantic               2.9.2
pydantic_core          2.23.4
pydantic-extra-types   2.9.0
pydantic-settings      2.6.1
Pygments               2.18.0
PyJWT                  2.9.0
PyMySQL                1.1.1
python-dotenv          1.0.1
python-multipart       0.0.17
PyYAML                 6.0.2
requests               2.32.3
rich                   13.9.3
shellingham            1.5.4
six                    1.16.0
sniffio                1.3.1
SQLAlchemy             2.0.36
sqlmodel               0.0.22
starlette              0.41.2
typer                  0.12.5
typing_extensions      4.12.2
ujson                  5.10.0
urllib3                2.2.3
uvicorn                0.32.0
watchfiles             0.24.0
websockets             13.1